test: run sdk server process in singleton mode by default

+++ eyes_server/applitools/eyes_server/server.py
@@ -20,7 +20,8 @@ executable_path = resource_filename(
 
 
 class SDKServer(object):
-    def __init__(self, executable=executable_path, singleton=False, log_file_name=None):
+    def __init__(self, singleton=False, log_file_name=None, executable=executable_path):
+        # type: (bool, str, str) -> None
         if log_file_name:
             self._log_file = open(log_file_name, "w+b")
             self.log_file_name = log_file_name
@@ -31,6 +32,8 @@ class SDKServer(object):
         command = [executable] if singleton else [executable, "--no-singleton"]
         self._sdk_process = Popen(command, stdout=self._log_file, stderr=STDOUT)
         self.port = self._read_port()
+        if singleton:
+            self._sdk_process = None  # leak the process, it self-terminates being idle
         self.is_closed = False  # explicit flag because python2 calls __del__ many times
         logger.info("Started Universal SDK server at %s", self.port)
 
@@ -47,7 +50,8 @@ class SDKServer(object):
         if not self.is_closed:
             self.is_closed = True
             logger.info("Quit Universal SDK server at %s", self.port)
-            self._sdk_process.terminate()
+            if self._sdk_process:
+                self._sdk_process.terminate()
             self._log_file.close()
 
     def __repr__(self):
